+++ jest.config.ts
@@ -3,7 +3,7 @@ import type { Config } from 'jest';
 const config: Config = {
 	preset: 'jest-preset-preact',
     testMatch: [ 
-        '**/__tests__/**/*.(js|ts|jsx|tsx)', 
+        '**/__tests__/**/*.test.(js|ts|jsx|tsx)', 
     ],
     extensionsToTreatAsEsm: ['.ts', '.tsx'],
     setupFilesAfterEnv: ['<rootDir>/__test__/config/setupTests.ts'],