+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### Added
 
 * add tests for UI components
-* add tests for Settings and Header
+* add tests for Settings, Summary, NoteList and Header
 * add Mock Service Worker library to simulate backend api in tests
 
 ## 1.2.0 - 28.09.2022